changed 1.2

+++ 09 Java/LABS/src/task01/task1_01_04.java
@@ -8,11 +8,11 @@ public class task1_01_04 {
         System.out.println("знаю");
         System.out.println("Java");
         // 2
-        System.out.println((46  +  10 ) * ( 10  /  3 ));
+        System.out.println((46  +  10 ) * ( 10  /  3d ));
         System.out.println((29 ) * ( 4 ) * ( -15 ));
         // 3
         int number = 10500;
-        int result = (number / 10) / 10;
+        double result = (number / 10d) / 10;
         System.out.println(result);
         // 4
         double result1 = 3.6 + 4.1 + 5.9; // переменную назвал result1, чтобы все идним разом посчитать
